+++ Tests/VCoreTests/Tests/Services and Managers/Network Client/NetworkClientTests.swift
@@ -27,9 +27,7 @@ final class NetworkClientTests: XCTestCase {
     
     // MARK: Data Tasks (Async)
     func testDataTaskNoDataAsync() async {
-        #if !os(watchOS)
-        guard NetworkReachabilityService.isConnectedToNetwork else { return }
-        #endif
+        guard NetworkReachabilityService.shared.isConnectedToNetwork else { return }
         
         do {
             var request: NetworkRequest = .init(url: "https://httpbin.org/post")
@@ -47,9 +45,7 @@ final class NetworkClientTests: XCTestCase {
     }
     
     func testDataTaskDataAsync() async {
-        #if !os(watchOS)
-        guard NetworkReachabilityService.isConnectedToNetwork else { return }
-        #endif
+        guard NetworkReachabilityService.shared.isConnectedToNetwork else { return }
         
         do {
             var request: NetworkRequest = .init(url: "https://httpbin.org/post")
@@ -68,9 +64,7 @@ final class NetworkClientTests: XCTestCase {
     }
     
     func testDataTaskJSONAsync() async {
-        #if !os(watchOS)
-        guard NetworkReachabilityService.isConnectedToNetwork else { return }
-        #endif
+        guard NetworkReachabilityService.shared.isConnectedToNetwork else { return }
         
         do {
             var request: NetworkRequest = .init(url: "https://httpbin.org/post")
@@ -88,9 +82,7 @@ final class NetworkClientTests: XCTestCase {
     }
     
     func testDataTaskJSONArrayAsync() async {
-        #if !os(watchOS)
-        guard NetworkReachabilityService.isConnectedToNetwork else { return }
-        #endif
+        guard NetworkReachabilityService.shared.isConnectedToNetwork else { return }
         
         do {
             var request: NetworkRequest = .init(url: "https://jsonplaceholder.typicode.com/posts")
@@ -107,9 +99,7 @@ final class NetworkClientTests: XCTestCase {
     }
     
     func testDataTaskDecodableAsync() async {
-        #if !os(watchOS)
-        guard NetworkReachabilityService.isConnectedToNetwork else { return }
-        #endif
+        guard NetworkReachabilityService.shared.isConnectedToNetwork else { return }
         
         do {
             var request: NetworkRequest = .init(url: "https://jsonplaceholder.typicode.com/posts")
@@ -127,9 +117,7 @@ final class NetworkClientTests: XCTestCase {
     
     // MARK: Data Tasks (Completion)
     func testDataTaskNoDataCompletion() {
-        #if !os(watchOS)
-        guard NetworkReachabilityService.isConnectedToNetwork else { return }
-        #endif
+        guard NetworkReachabilityService.shared.isConnectedToNetwork else { return }
         
         let expectation: XCTestExpectation = expectation(description: "NetworkClientCompletion")
         defer { waitForExpectations(timeout: 10) }
@@ -158,9 +146,7 @@ final class NetworkClientTests: XCTestCase {
     }
     
     func testDataTaskDataCompletion() {
-        #if !os(watchOS)
-        guard NetworkReachabilityService.isConnectedToNetwork else { return }
-        #endif
+        guard NetworkReachabilityService.shared.isConnectedToNetwork else { return }
         
         let expectation: XCTestExpectation = expectation(description: "NetworkClientCompletion")
         defer { waitForExpectations(timeout: 10) }
@@ -190,9 +176,7 @@ final class NetworkClientTests: XCTestCase {
     }
     
     func testDataTaskJSONCompletion() {
-        #if !os(watchOS)
-        guard NetworkReachabilityService.isConnectedToNetwork else { return }
-        #endif
+        guard NetworkReachabilityService.shared.isConnectedToNetwork else { return }
         
         let expectation: XCTestExpectation = expectation(description: "NetworkClientCompletion")
         defer { waitForExpectations(timeout: 10) }
@@ -221,9 +205,7 @@ final class NetworkClientTests: XCTestCase {
     }
 
     func testDataTaskJSONArrayCompletion() {
-        #if !os(watchOS)
-        guard NetworkReachabilityService.isConnectedToNetwork else { return }
-        #endif
+        guard NetworkReachabilityService.shared.isConnectedToNetwork else { return }
         
         let expectation: XCTestExpectation = expectation(description: "NetworkClientCompletion")
         defer { waitForExpectations(timeout: 10) }
@@ -251,9 +233,7 @@ final class NetworkClientTests: XCTestCase {
     }
 
     func testDataTaskDecodableCompletion() {
-        #if !os(watchOS)
-        guard NetworkReachabilityService.isConnectedToNetwork else { return }
-        #endif
+        guard NetworkReachabilityService.shared.isConnectedToNetwork else { return }
         
         let expectation: XCTestExpectation = expectation(description: "NetworkClientCompletion")
         defer { waitForExpectations(timeout: 10) }