+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+#### [3.5.0(27)]
+
+Services and Managers
+
+- `NetworkReachabilityService` now supports `watchOS`
+
 #### [3.4.0(25)](https://github.com/VakhoKontridze/VCore/releases/tag/3.4.0) — *2022 06 11*
 
 General